+++ packages/react-doctor/src/core/scan.ts
@@ -5,35 +5,35 @@ import {
   OXLINT_NODE_REQUIREMENT,
   PERFECT_SCORE,
   WARNING_RULE_PENALTY,
-} from "./constants.js";
+} from "../constants.js";
 import {
   printBrandingOnlyHeader,
   printScoreHeader,
   printNoScoreHeader,
-} from "./cli/render-score-header.js";
-import { printDiagnostics } from "./cli/render-diagnostics.js";
-import { printProjectDetection } from "./cli/render-project-detection.js";
-import { printSummary } from "./cli/render-summary.js";
-import { resolveOxlintNode } from "./cli/resolve-oxlint-node.js";
-import { NoReactDependencyError } from "./errors.js";
-import type { Diagnostic, ReactDoctorConfig, ScanOptions, ScanResult } from "./types.js";
+} from "../cli/render-score-header.js";
+import { printDiagnostics } from "../cli/render-diagnostics.js";
+import { printProjectDetection } from "../cli/render-project-detection.js";
+import { printSummary } from "../cli/render-summary.js";
+import { resolveOxlintNode } from "../cli/resolve-oxlint-node.js";
+import { NoReactDependencyError } from "../errors.js";
+import type { Diagnostic, ReactDoctorConfig, ScanOptions, ScanResult } from "../types.js";
 import {
   calculateScore,
   calculateScoreBreakdown,
   calculateScoreLocally,
-} from "./core/scoring/calculate-score.js";
-import { combineDiagnostics } from "./core/diagnostics/combine-diagnostics.js";
-import { discoverProject } from "./core/detection/discover-project.js";
-import { formatErrorChain } from "./cli/format-error-chain.js";
-import { highlighter } from "./cli/highlighter.js";
-import { computeJsxIncludePaths } from "./core/runners/jsx-include-paths.js";
-import { loadConfigWithSource } from "./core/config/load-config.js";
-import { isLoggerSilent, logger, setLoggerSilent } from "./cli/logger.js";
-import { resolveConfigRootDir } from "./core/config/resolve-config-root-dir.js";
-import { resolveLintIncludePaths } from "./core/runners/resolve-lint-include-paths.js";
-import { runKnip } from "./core/runners/run-knip.js";
-import { runOxlint } from "./core/runners/run-oxlint.js";
-import { isSpinnerSilent, setSpinnerSilent, spinner } from "./cli/spinner.js";
+} from "./scoring/calculate-score.js";
+import { combineDiagnostics } from "./diagnostics/combine-diagnostics.js";
+import { discoverProject } from "./detection/discover-project.js";
+import { formatErrorChain } from "../cli/format-error-chain.js";
+import { highlighter } from "../cli/highlighter.js";
+import { computeJsxIncludePaths } from "./runners/jsx-include-paths.js";
+import { loadConfigWithSource } from "./config/load-config.js";
+import { isLoggerSilent, logger, setLoggerSilent } from "../cli/logger.js";
+import { resolveConfigRootDir } from "./config/resolve-config-root-dir.js";
+import { resolveLintIncludePaths } from "./runners/resolve-lint-include-paths.js";
+import { runKnip } from "./runners/run-knip.js";
+import { runOxlint } from "./runners/run-oxlint.js";
+import { isSpinnerSilent, setSpinnerSilent, spinner } from "../cli/spinner.js";
 
 interface ResolvedScanOptions {
   lint: boolean;
