+++ packages/react-doctor/tests/install-skill.test.ts
@@ -2,7 +2,7 @@ import { existsSync, mkdirSync, mkdtempSync, rmSync, writeFileSync } from "node:
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vite-plus/test";
-import { runInstallSkill } from "../src/install-skill.js";
+import { runInstallSkill } from "../src/cli/install-skill.js";
 import { setLoggerSilent } from "../src/cli/logger.js";
 import { setSpinnerSilent } from "../src/cli/spinner.js";
